Relocate Playwright adapters for Twitter
